Moving all JavaScript into single 'main.js' file
On `bundle` the JS was split across multiple files, this was initially done as a SoC practice to ensure that work for certain components was easily tracked within Git (changes to Filter are transparent, for example); instead of all commits being made against one large main.js file, which could cause a lot of potential merge conflicts, etc.

Ideally we would look to keep that practice in place, but for now it may be more difficult to manage the multiple JS files with regards to the CMS integration; so as a temporary and quick solution we are going to move all JavaScript to one file and have just the one JS output.

In the future we hope to break away from a single large JS file, and split these out (as was being done), but we need to spend a bit of time looking at best practice and ways to concatenate all the individual JS files as part of the `bundle` task.